+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## master
 
+* Add support for property testing via testcheck-js.
+
 ## jest-cli 12.1.1
 
 * Windows stability fixes.
 * Mock module resolution fixes.
 * Remove test files from code coverage.
-* Add support for property testing via testcheck-js.
 
 ## jest-cli 12.1.0